Cleanup

+++ cpp/shared/s2p_util.cpp
@@ -243,8 +243,8 @@ string s2p_util::FormatSenseData(span<const byte> sense_data)
         return s;
     }
 
-    return s + fmt::format(", ILI: {0}, INFORMATION: {1}", static_cast<uint8_t>(sense_data[2]) & 0x20 ? "1" : "0", // NOSONAR Using byte type does not work with the bullseye compiler
-    GetInt32(sense_data, 3));
+    return s + fmt::format(", ILI: {0}, INFORMATION: {1}",
+        static_cast<uint8_t>(sense_data[2]) & static_cast<int>(0x20) ? "1" : "0", GetInt32(sense_data, 3));
 }
 
 string s2p_util::FormatSenseData(sense_key sense_key, asc asc, int ascq)
